#3767 misc changes
